requirements.txt
flask
pillow
numpy
cryptography
gunicorn
